Moved `ContentmentVersion` into "Configuration" namespace
